libscanR: Analysis and Visualization of LIBS Spectra

A vendor-agnostic pipeline for importing, preprocessing, calibrating, and visualizing Laser-Induced Breakdown Spectroscopy (LIBS) spectral data with a focus on biomedical tissue analysis.

Main functions

Import: ls_read_spectrum(), ls_read_dir(), ls_read_auto()

Preprocessing: ls_baseline(), ls_normalize(), ls_smooth(), ls_crop()

Peak analysis: ls_find_peaks(), ls_identify_peaks(), ls_element_db()

Calibration: ls_calibrate(), ls_saha_boltzmann(), ls_quantify()

Chemometrics: ls_pca(), ls_plsda(), ls_cluster()

Spatial mapping: ls_build_map(), ls_map_elements()

Tissue analysis: ls_tissue_classify(), ls_tissue_discriminate()

Interactive app: ls_run_app()

See also

Author

Maintainer: Raban Heller raban.heller@charite.de (ORCID)